Utilities/mount: Resolve regression on mounting non-storage-backed FSes
Without this patch, we fail on manually mounting RAMFS (which I tested
for) but any filesystem that is not backed by actual storage will fail.
This bug was introduced in 0739b5df11f2125f2f4b592e19c0d240620c5df6 and
now is resolved by checking if the source fd is negative, to avoid fail
of the fstat call on it.